Move HitSystem implementation to cpp file, close #177
